+++ Models/IdevsExportRequest.cs
@@ -96,13 +96,11 @@ public interface IIdevsExportRequest
     string? SelectionRange { get; set; }
     string? ConditionRange { get; set; }
     string? Logo { get; set; }
-    string? PageSize { get; set; }
-    PageMargin? Margin { get; set; }
+    PageSize PageSize { get; set; }
+    PageMargin Margin { get; set; }
     object? Entity { get; set; }
     TableTheme TableTheme { get; set; }
     IEnumerable<AggregateColumn>? AggregateColumns { get; set; }
-    string GetPageSize();
-    PageMargin GetMargin();
 }
 
 public class IdevsExportRequest : ListRequest, IIdevsExportRequest
@@ -113,12 +111,9 @@ public class IdevsExportRequest : ListRequest, IIdevsExportRequest
     public string? SelectionRange { get; set; }
     public string? ConditionRange { get; set; }
     public string? Logo { get; set; }
-    public string? PageSize { get; set; }
-    public PageMargin? Margin { get; set; }
+    public PageSize PageSize { get; set; } = new PageSize();
+    public PageMargin Margin { get; set; } = new PageMargin();
     public object? Entity { get; set; }
     public TableTheme TableTheme { get; set; }
     public IEnumerable<AggregateColumn>? AggregateColumns { get; set; }
-
-    public string GetPageSize() => string.IsNullOrWhiteSpace(PageSize) ? "A4 Landscape" : PageSize;
-    public PageMargin GetMargin() => Margin ?? new PageMargin();
 }